parser: fix parsing of constant
Constant type is unsigned long. However, convert to signed value in
print_constaant, as high enum value are much more likely intended to
be a negative signed value.

+++ objects.c
@@ -429,7 +429,7 @@ static pp_t print_base(obj_t *o, int depth, const char *prefix) {
 static pp_t print_constant(obj_t *o, int depth, const char *prefix) {
 	pp_t ret = {NULL, NULL};
 
-	asprintf_safe(&ret.prefix, "%s = %lu", o->name, o->constant);
+	asprintf_safe(&ret.prefix, "%s = %li", o->name, (long)o->constant);
 
 	return ret;
 }
